Split the test suite into several files
The test suite is a mess. We should improve this and the in a first
place, we should associate a file to a renderer's test case.
